# Potential perf improvements to result batching