I'm still new to Go and I don't really know what I'm doing but I'll add actual measurements soon.