@fbernier 2019-12-20
Benchmark needs cleanup but

```
+benchcmp master.bench current.bench
benchmark                 old ns/op     new ns/op     delta
BenchmarkBasicQuery-2     53447989      31190472      -41.64%
benchmark                 old allocs     new allocs     delta
BenchmarkBasicQuery-2     145639         138370         -4.99%
benchmark                 old bytes     new bytes     delta
BenchmarkBasicQuery-2     4491280       4392140       -2.21%
```