@fbernier 2020-01-10
@huebnerr same as #47